Add files via upload
The landslide detection system is successfully implemented as a prototype. All the sensors and other stuff works as per the expectations. The system design communication is via wifi, use of GSM is also implacable for such system. The precise threshold values for each of the sensors can be obtained through detailed study on the landslides and type of soil that the system is to be implemented in. Application of such system can also be developed in future by collaborating with IT department.